requirements.txt
absl-py
arrow
asgiref
asttokens
astunparse
beautifulsoup4
binaryornot
cachetools
certifi
chardet
charset-normalizer
click
colorama
comm
contourpy
cookiecutter
cycler
debugpy
decorator
django-browser-reload
django-tailwind
Django
djangorestframework
exceptiongroup
executing
filelock
flatbuffers
fonttools
fsspec
gast
gdown
google-auth-oauthlib
google-auth
google-pasta
grpcio
h5py
idna
imageio
importlib_resources
importlib-metadata
ipykernel
ipython
jedi
Jinja2
jupyter_client
jupyter_core
keras
kiwisolver
libclang
markdown-it-py
Markdown
MarkupSafe
matplotlib-inline
matplotlib
mdurl
mpmath
nest-asyncio
networkx
numpy
oauthlib
opt-einsum
packaging
parso
pillow
platformdirs
prompt-toolkit
protobuf
psutil
pure-eval
pyasn1-modules
pyasn1
Pygments
pyparsing
PySocks
python-dateutil
python-slugify
pytz
PyYAML
pyzmq
regex
requests-oauthlib
requests
rich
rsa
setuptools
six
soupsieve
sqlparse
stack-data
sympy
tensorboard-data-server
tensorboard
tensorflow-estimator
tensorflow-io-gcs-filesystem
tensorflow
termcolor
text-unidecode
torch
torchvision
tornado
tqdm
traitlets
typer
types-python-dateutil
typing_extensions
tzdata
urllib3
wasabi
wcwidth
weasel
Werkzeug
wheel
wrapt
xgboost
zipp
